server/pyproject.toml
[project]
name = "pidentify-server"
version = "0.1.0"
description = ""
authors = [
    {name = "Pat Dill",email = "ppygrl@proton.me"}
]
readme = "README.md"
requires-python = ">=3.11,<4.0"
dependencies = [
    "sounddevice (==0.5.1)",
    "soundfile (==0.13.0)",
    "numpy (>=2.3.1,<3.0.0)",
    "asyncclick (>=8.1.8,<9.0.0)",
    "pydantic (>=2.11.7,<3.0.0)",
    "cachetools (>=6.1.0,<7.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "fastapi[standard] (>=0.116.1,<0.117.0)",
    "alembic (>=1.16.4,<2.0.0)",
    "alembic-utils (>=0.8.8,<0.9.0)",
    "sqlalchemy (>=2.0.41,<3.0.0)",
    "sqlalchemy-utils (>=0.41.2,<0.42.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "mutagen (>=1.47.0,<2.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "sqlalchemy-utc (>=0.14.0,<0.15.0)",
    "pyyaml (>=6.0.3,<7.0.0)",
    "argon2-cffi (>=25.1.0,<26.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "platformdirs (>=4.0.0,<5.0.0)",
    "pyzmq (>=26.0.0,<27.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false
